Migrate from 1.18 to 1.19
Added LevelRoomDetector

This class can be used to distribute the behaviour from
`LevelRoomDetection.findRooms` across multiple ticks.

Fixed potential lockup

The lockup occurs when a room scan is done on a room with no doors.

Roll overlap detection into LevelRoomDetector

Add Debug Art for debugging and bug reporting

Fixing crash

Improving debug art logic

Added regression test

WIP: "Flight Recorder"

to make it easier to understand the process that was used without having
to use the debugger.

"Flight Recorder"

to make it easier to understand the process that was used without having
to use the debugger.